Create Sub Percentage_Change().vb
The percentage change works. But I have a bug associated with a counter insider a nested loop: "Ticker_row_start = Ticker_row_start +1".
It creates a ""Run-time error '6': Overflow" error.
The error doesn't happen if I run the two loops consecutively (and not as one loop nested inside the other). But then my values are wrong.