Deleted some early drafts
They were very preliminary. Most the work that was done in them is saved in previous version of my final script.

Early on I was creating and saving new files as I went to keep track of my progress. Then I realised I didn't need to and all the files were getting messy.